Use rawText and Prevent Errors
Add post rawText to the database instead of text. Filter out invalid file name characters to prevent filename errors. These changes allow the {rawText} tag to be used for PostFileNameFormat and PaidPostFileNameFormat. {text} tag can be used for the others (as well as the previously mentioned ones) though html content may be present in the output.