Reverse Lists Before Processing
The reasoning for this change is that creators can repost content that is already in their media library. When traversing the lists in forward order, the most recent posts are taken and any previous duplicates are skipped. I think it more desirable to have the original post date (and original post text) associated with the downloaded media, not the reposted date and text.